Restructure kengdic_2011
Rewrite `kengdic_2011.tsv` to `kengdic.tsv`. Restruture the fields as
follows:

* Level: remove the `mt` prefix and set `t` values to NULL. The
documentation from Joe said that the values were only supposed to be A,
B, C or D, so it looks like the field were garbled somehow.

* pos, posn: remove these fields entirely. The documentation said that
the fields had been largely neglected, and it shows in the data, where
appears that the entire file was set to a single part of speech.

* syn: this field was blank everywhere, and the purpose was not
documented, remove it entirely.

* wordsize: this contains neither the number of characters, nor the
number of individual hangeul components (by any counting scheme that I
can devise). It's not documented, and any size-related information
could be re-derived anyway, so remove it entirely.

* Create a new source field. The format is
`submitter-topic@ezcorean:old_ID`. Moving forward, the submitter name
and topic IDs are rather specific to the old method of entry creation
via ezcorean, and a `source` field naturally submsumes them. Note that
the topic ID was sometimes the same as the word ID, but not always.

* wordid: replace with a new, sequentially-assigned ID. The old ID is
still there in the `source` field. I decided to replace the ID when I
realized that they weren't unique! The following IDs were duplicated:

	235658
	235655
	222666
	230612
	230613
	232791
	235656
	235657
	230057
	230518

Fixes #4.

Script used to do the rewriting:

	from collections import defaultdict
	import csv
	from itertools import chain
	from pathlib import Path
	import sys

	project_dir = Path(__name__).parent

	input_header = (
	    "id\tword\tsyn\tdef\tposn\tpos\tsubmitter\tdoe\tword_size\thanja\ttopic_id\tlevel\n"
	)
	output_header = "id\tsurface\thanja\tgloss\tlevel\tcreated\tsource"

	def main():
	    print(output_header)
	    seen_ids = set()
	    word_freq = defaultdict(int)
	    with open(project_dir / "kengdic_2011.tsv") as f:
	        id_counter = 1
	        rows = csv.DictReader(chain([input_header], f), delimiter="\t")
	        for r in rows:
	            r = {k: v.strip() for k, v in r.items()}

	            word_freq[r["word"]] += 1
	            if r["id"] in seen_ids:
	                print(f"WARNING: duplicate ID {r['id']}", file=sys.stderr)
	            seen_ids.add(r["id"])

	            # current IDs are not unique and likely not needed, so assign new ones
	            r["id_counter"] = id_counter
	            id_counter += 1
	            # clean up weirdly-formatted field
	            if r["level"].startswith("mt"):
	                r["level"] = r["level"][2]
	            else:
	                r["level"] = None
	            # the dictionary will grow beyond ezcorean data, so
	            # generalize the idea of a submitter to a source
	            r["source"] = "{submitter}-{topic_id}@ezcorean:{id}".format(**r)

	            print("{id_counter}\t{word}\t{hanja}\t{def}\t{level}\t{doe}\t{source}".format(**r))

	    with open(project_dir / 'reports' / 'duplicates.txt', 'w') as f:
	        f.write("Word\toccurrences\n")
	        for word, freq in word_freq.items():
	            if freq > 1:
	                f.write(f"{word}\t{freq}\n")

	if __name__ == "__main__":
	    main()