Convert timestamps to UTC and remove NULL/None
While working on a datapackage file, I found that the standard/default
timestamp format is ISO 8601. Assume that the current timestamps are in
the KST timezone, then rewrite them in ISO 8601 with the UTC timezone.
Note that, for an unknown reason, the timezone printed from the
`isoformat` method is not correctly formatted, so we have to replace it
ourselves.

Remove `None` and `NULL` string values, replacing them with more general
empty strings.

The entry for item 39679 did not have a `created` timestamp, so I've put
in an approximate one by looking at other timestamps from the same user.

Below is the python script used to process the file:

    import csv
    import sys

    from dateutil.parser import parse
    from dateutil import tz

    UTC = tz.gettz("UTC")

    with open("kengdic.tsv") as f:
        rows = csv.DictReader(f, delimiter="\t")
        # re-print header
        print("id\tsurface\thanja\tgloss\tlevel\tcreated\tsource")
        for row in rows:
            # Convert date from KST to UTC, then rewrite in ISO 8601 format
            if row["created"] != "NULL":
                created = row["created"] + " UTC+9"
                stamp = parse(created).astimezone(UTC)
                row["created"] = stamp.isoformat().replace("+00:00", "Z")
            else:
                print(f"Missing timestamp for {row['id']}", file=sys.stderr)

            # Replace SQL- and python-specific missing values with generic empty values
            for key, val in row.items():
                if val in ['NULL', 'None']:
                    row[key] = ''

            print("{id}\t{surface}\t{hanja}\t{gloss}\t{level}\t{created}\t{source}".format(**row))