Cleanups focused on matching hanja field
With the goal to make hanja fields the same length as the surface field,
perform thousands of cleanups:

* Replace all dashes in hanja fields with the hangeul or spaces, etc.
that would be used when spelling the word using hanja. We should still
do a sanity check to make sure that the hanja and surface fields are
the same length (there are probably many mismatching cases).
* Add or edit English, surface, and hanja fields where issues are
noticed and the fixes are clear from investigating various dictionaries
and google.

Also remove a few of the non-sensical entries and add a few obvious ones
at the end.